refactor(spi): move StatusTransitionPolicy into com.embabel.dice.spi
StatusTransitionPolicy is a policy SPI ("consumers may supply their own
implementation") that was left in common while the other policy SPIs moved to
spi — a split-brain a reviewer flagged. Move it and its shipped default
DecayStatusPolicy (same file, no other-dev code involved) into spi. It imports
only proposition types, so spi stays a clean base layer.

Also bring this branch's AGENTS.md package maps current with the earlier spi
extraction (they still listed the trust/conflict types under common/revision).
